requirements.txt
flask
requests
python-dotenv
bcrypt